+++ modules/services/compose/src/main/java/au/com/shiftyjelly/pocketcasts/compose/buttons/RowOutlinedButton.kt
@@ -5,11 +5,17 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.ButtonColors
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.OutlinedButton
 import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Share
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
@@ -25,7 +31,10 @@ fun RowOutlinedButton(
     text: String,
     modifier: Modifier = Modifier,
     includePadding: Boolean = true,
-    onClick: () -> Unit
+    border: BorderStroke = outlinedBorder,
+    colors: ButtonColors = ButtonDefaults.outlinedButtonColors(),
+    iconImage: ImageVector? = null,
+    onClick: () -> Unit,
 ) {
     Row(
         modifier = modifier
@@ -35,9 +44,15 @@ fun RowOutlinedButton(
         OutlinedButton(
             onClick = { onClick() },
             shape = RoundedCornerShape(12.dp),
-            border = outlinedBorder,
-            modifier = Modifier.fillMaxWidth()
+            border = border,
+            colors = colors,
+            modifier = Modifier
+                .fillMaxWidth()
         ) {
+            iconImage?.let {
+                Icon(imageVector = it, contentDescription = "")
+            }
+
             Text(
                 text = text,
                 fontSize = 18.sp,
@@ -52,7 +67,8 @@ fun RowOutlinedButton(
 fun RowOutlinedButtonLightPreview() {
     AppTheme(Theme.ThemeType.LIGHT) {
         RowOutlinedButton(
-            text = "Accept",
+            text = "Share",
+            iconImage = Icons.Default.Share,
             onClick = {}
         )
     }
@@ -63,7 +79,8 @@ fun RowOutlinedButtonLightPreview() {
 fun RowOutlinedButtonDarkPreview() {
     AppTheme(Theme.ThemeType.DARK) {
         RowOutlinedButton(
-            text = "Accept",
+            text = "Share",
+            iconImage = Icons.Default.Share,
             onClick = {}
         )
     }